refactor: move simple-pricing to Node plugins
