+++ imports/node-app/plugins/simple-pricing/queries/index.js
@@ -1,0 +1,5 @@
+import getVariantPrice from "./getVariantPrice.js";
+
+export default {
+  getVariantPrice
+};
