+++ imports/node-app/plugins/simple-pricing/resolvers/CatalogProduct.js
@@ -1,4 +1,4 @@
-import xformPricingArray from "../util/xformPricingArray";
+import xformPricingArray from "../util/xformPricingArray.js";
 
 export default {
   pricing: (node) => xformPricingArray(node.pricing)
