+++ imports/node-app/plugins/simple-pricing/resolvers/ProductPricingInfo.js
@@ -1,8 +1,8 @@
-import { getXformedCurrencyByCode } from "@reactioncommerce/reaction-graphql-xforms/currency";
-import xformCurrencyExchangePricing from "../util/xformCurrencyExchangePricing";
+import getCurrencyDefinitionByCode from "@reactioncommerce/api-utils/getCurrencyDefinitionByCode.js";
+import xformCurrencyExchangePricing from "../util/xformCurrencyExchangePricing.js";
 
 export default {
-  currency: (node) => getXformedCurrencyByCode(node.currencyCode),
+  currency: (node) => getCurrencyDefinitionByCode(node.currencyCode),
   currencyExchangePricing: (node, { currencyCode }, context) => xformCurrencyExchangePricing(node, currencyCode, context),
   compareAtPrice: ({ compareAtPrice: amount, currencyCode }) => (typeof amount === "number" && amount > 0 ? { amount, currencyCode } : null)
 };
