+++ imports/node-app/plugins/simple-pricing/resolvers/index.js
@@ -1,0 +1,9 @@
+import CatalogProduct from "./CatalogProduct.js";
+import CatalogProductVariant from "./CatalogProductVariant.js";
+import ProductPricingInfo from "./ProductPricingInfo.js";
+
+export default {
+  CatalogProduct,
+  CatalogProductVariant,
+  ProductPricingInfo
+};
