+++ imports/node-app/plugins/simple-pricing/util/getDisplayPrice.test.js
@@ -1,4 +1,4 @@
-import getDisplayPrice from "./getDisplayPrice";
+import getDisplayPrice from "./getDisplayPrice.js";
 
 const mockUSDCurrencyInfo = {
   enabled: true,
