+++ imports/node-app/plugins/simple-pricing/util/getVariantPriceRange.js
@@ -1,4 +1,4 @@
-import getPriceRange from "./getPriceRange";
+import getPriceRange from "./getPriceRange.js";
 
 /**
  *
