+++ imports/node-app/plugins/simple-pricing/util/mutateNewProductBeforeCreate.js
