+++ imports/plugins/core/accounts/server/no-meteor/resolvers/Account/index.js
@@ -1,13 +1,13 @@
 import { get } from "lodash";
 import { encodeAccountOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/account";
-import { getXformedCurrencyByCode } from "@reactioncommerce/reaction-graphql-xforms/currency";
+import getCurrencyDefinitionByCode from "@reactioncommerce/api-utils/getCurrencyDefinitionByCode.js";
 import { resolveShopFromShopId } from "@reactioncommerce/reaction-graphql-utils";
 import addressBook from "./addressBook";
 
 export default {
   _id: (account) => encodeAccountOpaqueId(account._id),
   addressBook,
-  currency: (account) => getXformedCurrencyByCode(account.profile && account.profile.currency),
+  currency: (account) => getCurrencyDefinitionByCode(account.profile && account.profile.currency),
   emailRecords: (account) => account.emails,
   firstName: (account) => account.profile.firstName,
   lastName: (account) => account.profile.lastName,
