+++ imports/plugins/core/shop/server/resolvers/Shop/index.js
@@ -1,4 +1,4 @@
-import { getXformedCurrencyByCode, getXformedCurrenciesByShop } from "@reactioncommerce/reaction-graphql-xforms/currency";
+import getCurrencyDefinitionByCode from "@reactioncommerce/api-utils/getCurrencyDefinitionByCode.js";
 import { encodeNavigationTreeOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/navigationTree";
 import { encodeShopOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/shop";
 import brandAssets from "./brandAssets";
@@ -8,8 +8,17 @@ import tags from "./tags";
 export default {
   _id: (node) => encodeShopOpaqueId(node._id),
   brandAssets,
-  currencies: (shop) => getXformedCurrenciesByShop(shop),
-  currency: (shop) => getXformedCurrencyByCode(shop.currency),
+  currencies(shop) {
+    if (!shop || !shop.currencies) return [];
+
+    // map over all provided currencies, provided in the format stored in our Shops collection,
+    // and convert them to the format that GraphQL needs
+    return Object.keys(shop.currencies).map((code) => ({
+      ...getCurrencyDefinitionByCode(code),
+      enabled: shop.currencies[code].enabled || false
+    }));
+  },
+  currency: (shop) => getCurrencyDefinitionByCode(shop.currency),
   defaultNavigationTreeId: (shop) => encodeNavigationTreeOpaqueId(shop.defaultNavigationTreeId),
   defaultNavigationTree,
   tags
