--- imports/plugins/included/simple-pricing/server/no-meteor/queries/index.js
@@ -1,5 +1,0 @@
-import getVariantPrice from "./getVariantPrice";
-
-export default {
-  getVariantPrice
-};
