--- imports/plugins/included/simple-pricing/server/no-meteor/resolvers/index.js
@@ -1,9 +1,0 @@
-import CatalogProduct from "./CatalogProduct";
-import CatalogProductVariant from "./CatalogProductVariant";
-import ProductPricingInfo from "./ProductPricingInfo";
-
-export default {
-  CatalogProduct,
-  CatalogProductVariant,
-  ProductPricingInfo
-};
